+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/common/repository/CertificatesHolder.kt
@@ -1,0 +1,7 @@
+package de.rki.coronawarnapp.covidcertificate.common.repository
+
+import de.rki.coronawarnapp.covidcertificate.common.certificate.CwaCovidCertificate
+
+interface CertificatesHolder {
+    val allCertificates: Set<CwaCovidCertificate>
+}
