+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/recovery/core/RecoveryCertificatesHolder.kt
@@ -1,0 +1,12 @@
+package de.rki.coronawarnapp.covidcertificate.recovery.core
+
+import de.rki.coronawarnapp.covidcertificate.common.repository.CertificatesHolder
+
+data class RecoveryCertificatesHolder(
+    val certificates: Set<RecoveryCertificateWrapper>,
+    val recycledCertificates: Set<RecoveryCertificate>
+) : CertificatesHolder {
+    override val allCertificates: Set<RecoveryCertificate> by lazy {
+        certificates.map { it.recoveryCertificate }.toSet() + recycledCertificates
+    }
+}
